requirements.txt
pennylane>=0.40.0
pennylane-lightning>=0.40.0
torch>=2.2.0
torchvision>=0.17.0
numpy>=1.26.0
scipy>=1.12.0
scikit-learn>=1.4.0
pandas>=2.2.0
pyyaml>=6.0.1
matplotlib>=3.8.0
tqdm>=4.66.0
codecarbon>=2.4.0
